Changed move integers to bytes.

+++ src/PokeGame.Infrastructure/Converters/PowerConverter.cs
@@ -6,7 +6,7 @@ internal class PowerConverter : JsonConverter<Power>
 {
   public override Power? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetInt32(out int value) ? new Power(value) : null;
+    return reader.TryGetByte(out byte value) ? new Power(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, Power power, JsonSerializerOptions options)
